DBHelper updates Count and LastTry

+++ app/src/main/java/com/timkatimkah/vocaboptimal/VocabEntry.java
@@ -50,7 +50,7 @@ class VocabEntry {
         return lastTry;
     }
 
-    void setLastTry(int lastTry) {
+    void setLastTry(long lastTry) {
         this.lastTry = lastTry;
     }
 }
